Change Dto structure, rename Dtos
